move packet attributes to submodule
